requirements.txt
environs==9.5.0
redis==4.5.5
python-telegram-bot==13.15
vk_api==11.9.9